feat: add validation for zero VAT rate and enhance receipt validation with reference checks
